Move classes from UserDto into separate files
