+++ src/main/scala/com/softwaremill/realworld/users/UsersEndpoints.scala
@@ -3,6 +3,7 @@ package com.softwaremill.realworld.users
 import com.softwaremill.realworld.common.*
 import com.softwaremill.realworld.db.{Db, DbConfig}
 import com.softwaremill.realworld.http.ErrorMapper.defaultErrorsMappings
+import com.softwaremill.realworld.users.model.{User, UserRegister, UserUpdate, UserLogin}
 import io.getquill.SnakeCase
 import sttp.model.StatusCode
 import sttp.tapir.generic.auto.*
