+++ src/main/scala/com/softwaremill/realworld/users/model/UserData.scala
@@ -1,0 +1,15 @@
+package com.softwaremill.realworld.users.model
+
+import zio.json.{DeriveJsonDecoder, DeriveJsonEncoder}
+
+object UserData:
+  given userDataEncoder: zio.json.JsonEncoder[UserData] = DeriveJsonEncoder.gen[UserData]
+  given userDataDecoder: zio.json.JsonDecoder[UserData] = DeriveJsonDecoder.gen[UserData]
+
+case class UserData(
+    email: String,
+    token: Option[String],
+    username: String,
+    bio: Option[String],
+    image: Option[String]
+)
