+++ src/main/scala/com/softwaremill/realworld/users/model/UserLoginData.scala
@@ -1,0 +1,12 @@
+package com.softwaremill.realworld.users.model
+
+import zio.json.{DeriveJsonDecoder, DeriveJsonEncoder}
+
+object UserLoginData:
+  given userLoginDataEncoder: zio.json.JsonEncoder[UserLoginData] = DeriveJsonEncoder.gen[UserLoginData]
+  given userLoginDataDecoder: zio.json.JsonDecoder[UserLoginData] = DeriveJsonDecoder.gen[UserLoginData]
+
+case class UserLoginData(
+    email: String,
+    password: String
+)
