+++ src/main/scala/com/softwaremill/realworld/users/model/UserRegisterData.scala
@@ -1,0 +1,13 @@
+package com.softwaremill.realworld.users.model
+
+import zio.json.{DeriveJsonDecoder, DeriveJsonEncoder}
+
+object UserRegisterData:
+  given userRegisterDataEncoder: zio.json.JsonEncoder[UserRegisterData] = DeriveJsonEncoder.gen[UserRegisterData]
+  given userRegisterDataDecoder: zio.json.JsonDecoder[UserRegisterData] = DeriveJsonDecoder.gen[UserRegisterData]
+
+case class UserRegisterData(
+    email: String,
+    username: String,
+    password: String
+)
