+++ src/main/scala/com/softwaremill/realworld/users/model/UserRow.scala
@@ -1,0 +1,10 @@
+package com.softwaremill.realworld.users.model
+
+case class UserRow(
+    userId: Int,
+    email: String,
+    username: String,
+    password: String,
+    bio: Option[String],
+    image: Option[String]
+)
