+++ src/main/scala/com/softwaremill/realworld/users/model/UserSession.scala
@@ -1,0 +1,3 @@
+package com.softwaremill.realworld.users.model
+
+case class UserSession(email: String)
