+++ src/main/scala/com/softwaremill/realworld/users/model/UserWithPassword.scala
@@ -1,0 +1,6 @@
+package com.softwaremill.realworld.users.model
+
+case class UserWithPassword(
+    user: UserData,
+    hashedPassword: String
+)
