+++ src/test/scala/com/softwaremill/realworld/common/JsonEncodingSpec.scala
@@ -2,7 +2,8 @@ package com.softwaremill.realworld.common
 import com.softwaremill.realworld.articles.ArticlesEndpointsSpec.{suite, test}
 import com.softwaremill.realworld.articles.model.{Article, ArticleAuthor, ArticleData}
 import com.softwaremill.realworld.articles.model
-import com.softwaremill.realworld.users.{User, UserData}
+import com.softwaremill.realworld.users
+import com.softwaremill.realworld.users.model.{User, UserData}
 import zio.json.*
 import zio.test.*
 import zio.test.Assertion.{equalTo, isEmpty, isNegative}
@@ -14,7 +15,7 @@ object JsonEncodingSpec extends ZIOSpecDefault {
   override def spec = suite("JSON encoding for data objects") {
     suite("User related objects")(
       test("User fields with None value are present in rendered json as null values") {
-        val user: User = User(
+        val user: User = users.model.User(
           UserData(
             email = "email@domain.com",
             token = None,
